+++ src/DagScheduler.jl
@@ -23,7 +23,7 @@ const META_IMPL = Dict(
 include("common.jl")
 include("bcast_channel.jl")
 include("meta_stores/meta_store.jl")
-include("scheduler.jl")
+include("scheduling/scheduler.jl")
 include("execution/queue.jl")
 include("execution/engine.jl")
