+++ src/common.jl
@@ -1,3 +1,6 @@
+#------------------------------------
+# Common Types
+#------------------------------------
 const TaskIdType = UInt64
 
 const NoTask = TaskIdType(0)
@@ -6,6 +9,72 @@ taskid(th::Thunk) = TaskIdType(th.id)
 #taskid(ch::Chunk) = TaskIdType(hash(ch))
 #taskid(executable) = TaskIdType(hash(executable))
 
+mutable struct NodeEnv
+    brokerid::UInt64
+    broker_task::Union{Future,Void}
+    executorids::Vector{UInt64}
+    executor_tasks::Vector{Future}
+    last_task_stat::Vector{Tuple{UInt64,Future}}
+
+    function NodeEnv(brokerid::Integer, executorids::Vector{Int})
+        new(brokerid, nothing, executorids, Vector{Future}(), Vector{Tuple{UInt64,Future}}())
+    end
+end
+
+mutable struct RunEnv
+    rootpath::String
+    masterid::UInt64
+    nodes::Vector{NodeEnv}
+    reset_task::Union{Task,Void}
+    debug::Bool
+
+    function RunEnv(; rootpath::String="/dagscheduler", masterid::Int=myid(), nodes::Vector{NodeEnv}=[NodeEnv(masterid,workers())], debug::Bool=false)
+        nexecutors = 0
+        for node in nodes
+            nw = length(node.executorids)
+            (nw > 1) || error("need at least two workers on each node")
+            nexecutors += nw
+        end
+        (nexecutors > 1) || error("need at least two workers")
+
+        # ensure clean paths
+        delete_meta(UInt64(masterid), nodes, rootpath)
+
+        @everywhere MemPool.enable_who_has_read[] = false
+        @everywhere Dagger.use_shared_array[] = false
+
+        new(rootpath, masterid, nodes, nothing, debug)
+    end
+end
+
+struct ComputeCost
+    cpu::Float64        # cost of computing
+    net::Float64        # cost of transfering data
+end
+
+fval(c::ComputeCost) = (c.cpu + c.net)
+
+function +(c1::ComputeCost, c2::ComputeCost)
+    ComputeCost(c1.cpu + c2.cpu, c1.net + c2.net)
+end
++(c1::ComputeCost) = c1
+isless(c1::ComputeCost, c2::ComputeCost) = fval(c1) < fval(c2)
+/(c1::ComputeCost, c2::ComputeCost) = fval(c1) / fval(c2)
+
+# here is where the network cost can be adjusted to account for the computation that this Thunk does
+# if it is a reduction step, apply the reduction before summing the input costs
+cost_function(t::Thunk) = sum
+
+# consider splitting node for scheduling only if there is no clear association with any node
+const SPLIT_COST_RATIO = 1.5
+# consider splitting node for scheduling only if the compute cost is substantial
+const SPLIT_CAPACITY_RATIO = 0.5
+
+mean_node_capacity(runenv::RunEnv) = mean([length(node.executorids)*1 for node in runenv.nodes])
+
+#------------------------------------
+# Logging utilities
+#------------------------------------
 function tasklog(env, msg...)
     env.debug && info(env.role, " : ", env.id, " : ", env.brokerid, " : ", msg...)
 end
@@ -17,6 +86,9 @@ function taskexception(env, ex, bt)
     xret
 end
 
+#------------------------------------
+# DAG Utilities
+#------------------------------------
 function collect_chunks(dag)
     if isa(dag, Chunk)
         return collect(dag)
@@ -101,19 +173,4 @@ dref_to_fref!(dag) = walk_dag(dag, true) do node,depth
     else
         node
     end
-end
-
-execution_stages(dag_node) = execution_stages!(deepcopy(dag_node))
-function execution_stages!(dag_node, deps=Dagger.dependents(dag_node), root=dag_node)
-    newinps = Set{Thunk}()
-    for inp in dag_node.inputs
-        if istask(inp)
-            for cinp in execution_stages!(inp, deps, root)
-                push!(newinps, cinp)
-            end
-        end
-    end
-    dag_node.inputs = tuple(newinps...)
-    ((dag_node == root) || (length(deps[dag_node]) > 1)) && (return [dag_node])
-    [dag_node.inputs...]
 end
